+++ app/src/main/java/org/linphone/activities/main/files/fragments/VideoViewerFragment.kt
@@ -26,8 +26,10 @@ import android.widget.MediaController
 import androidx.lifecycle.ViewModelProvider
 import androidx.navigation.fragment.findNavController
 import org.linphone.R
+import org.linphone.activities.main.MainActivity
 import org.linphone.activities.main.files.viewmodels.VideoFileViewModel
 import org.linphone.activities.main.files.viewmodels.VideoFileViewModelFactory
+import org.linphone.compatibility.Compatibility
 import org.linphone.core.tools.Log
 import org.linphone.databinding.FileVideoViewerFragmentBinding
 
@@ -67,6 +69,13 @@ class VideoViewerFragment : GenericViewerFragment<FileVideoViewerFragmentBinding
             }
         }
         initMediaController()
+
+        viewModel.fullScreenMode.observe(
+            viewLifecycleOwner
+        ) { hide ->
+            Compatibility.hideAndroidSystemUI(hide, requireActivity().window)
+            (requireActivity() as MainActivity).hideStatusFragment(hide)
+        }
     }
 
     override fun onResume() {